Moved HashChkParser and Output classes to their own file, hashchk_parser.py for organizational purposes
